refactor: Prisma 클라이언트 및 타입 경로 수정
