+++ src/lib/prisma.ts
@@ -1,3 +1,3 @@
-import { PrismaClient } from '../generated/prisma';
+import { PrismaClient } from '@prisma/client';
 
 export const prisma = new PrismaClient();
