+++ src/types/express.d.ts
@@ -1,4 +1,4 @@
-import { User } from '../generated/prisma';
+import { User } from '@prisma/client';
 import type { OffsetQueryDto, CursorQueryDto } from './queryTypes';
 
 declare global {
