+++ src/types/userTypes.ts
@@ -4,13 +4,14 @@ export interface User {
   email: string;
   nickname: string;
   password: string;
+  refreshToken: string | null;
   image: string | null;
   createdAt: Date;
   updatedAt: Date;
 }
 
 // DTO
-export type UserResponseDTO = Omit<User, 'password'>;
+export type UserResponseDTO = Omit<User, 'password' | 'refreshToken'>;
 
 export interface UpdateUserDTO {
   nickname?: string;
